GQL request-response roundtrip #30
* an untidy, but functional version
* replaced usize with u64 thruout

+++ stm_html_ui/src/html/html_data.rs
@@ -123,5 +123,5 @@ impl KeywordMetadata {
 #[derive(Serialize, Deserialize, Debug)]
 pub(crate) struct RelatedKeywords {
     pub k: String,
-    pub c: usize,
+    pub c: u64,
 }
